V2.0.1 -- 20150105
修改对是否选课成功的验证方式，防止由于选课系统返回空内容导致的误判
以及。。改了一下目录的名字。。
Mac下的可执行文件已更新
