--- NK-CourseGrabber/Source_Code/140710dev/vcodeData2
@@ -1,10 +1,0 @@
-000111111000011000000100010000000010100000000001100000000001100000000001010000000010011000000110000111111000
-010000000001100000000001111111111111000000000001000000000001
-0110101010100100
-01101010101001
-000000011000000000101000000001001000000110001000001000001000110000001000111111111111000000001000000000001000
-1111110100001010000101000010100000110000011000000
-000111111100001000100010010001000001100001000001100001000001100001000001100000100010010000011100
-11111111
-000000011100011100100010100011000001100001000001100001000001100001000001100011000001011100100010000000011100
-00111000100010100000110000011000001100000101000100011111
