+++ NK-CourseGrabber/Source_Code/Dev/vcodeData
@@ -1,0 +1,10 @@
+00011111000011000001101000000000110000000001100000000010110000011000011111000
+01000000000010000000011111111111100000000001
+01100000011100000001011000000100110000010001110011000010111000000100000000010
+011010101101
+0000001100000000101000000010010000011000100001000001000111111111110000000100000000001000
+000100111010010100101000110000
+00001111100000101000100010100000101001000001010010000011000010001010000011100
+01101010101011
+00110001100010010100101000010000110000100001100001100010100101001000110001100
+00111000001010001000011000001000110000010010100000101000100001100000111100000
