--- NK-CourseGrabber/Update_Log.txt
@@ -1,13 +1,0 @@
-﻿[20140711 07:46	V2.0]
-	这个版本最大的不同就是支持12门课分批提交╮(￣▽￣")╭ 算是我本来的计划里面想添加的最后一个功能了喵
-	让我们一起膜拜@白银硕 大神
-	程序喵
-[20140707 17:12	V1.3]
-	喵，请不要在意程序喵写代码时候留下的各种Bug……
-	程序喵
-[20140703 16:55	V1.2]
-	喵，本次更新的原因是…………日志太长会把内存玩儿坏……喵……
-	程序喵
-[20140702 14:58	V1.1]
-	喵，本次更新的原因是因为没有考虑到如果选课系统没开会怎样，所以加了几行代码
-	程序喵
